requirements.txt
django==4.2
loguru==0.7.0
asyncio==3.4.3
requests==2.25.1
beautifulsoup4==4.11.1
free-proxy-list==0.0.3
random-user-agent==1.0.1
lxml==4.8.0
openpyxl==3.0.10
celery==5.2.7